requirements.txt
numpy
matplotlib
scipy
Shapely
pygame
networkx
pyyaml